Fixed all the things

+++ 8BitEngine/src/LuaComponent.cpp
@@ -19,7 +19,7 @@ LuaComponent::LuaComponent(const std::string& fileName, const std::string& varia
 	RegisterFunctions();
 
 	// Set variable
-	m_Lua.SetGlobalNumber(variable, number);
+	m_Lua.SetGlobalNumber(variable, (float) number);
 
 	// Load script into state
 	m_Lua.ExecuteFile(fileName);
